requirements.txt
Flask
pymongo
Flask-JWT-Extended
marshmallow
Werkzeug
python-dotenv
bcrypt
gunicorn
eventlet
pendulum
bleach